chore(uptime): Move UptimeDomainCheckFailure to uptime/grouptype.py take 2
We want to keep grouptypes in the module that is related to them, so just moving this out.

This caused pickling problems last time, so importing the model in its old location to prevent issues.
